# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Actor": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "title": "Actorid",
            "type": "string"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "role": {
            "enum": [
              "NURSE",
              "CLINICIAN"
            ],
            "title": "Role",
            "type": "string"
          },
          "tenantId": {
            "title": "Tenantid",
            "type": "string"
          }
        },
        "required": [
          "actorId",
          "displayName",
          "tenantId",
          "role"
        ],
        "title": "Actor",
        "type": "object"
      },
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlationId": {
            "title": "Correlationid",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "BiologicalSex": {
        "enum": [
          "FEMALE",
          "MALE",
          "OTHER",
          "UNKNOWN"
        ],
        "title": "BiologicalSex",
        "type": "string"
      },
...
      "DemoLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Actorid",
            "type": "string"
          }
        },
        "required": [
          "actorId"
        ],
        "title": "DemoLoginRequest",
        "type": "object"
      },
      "DemoLoginResponse": {
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "tokenType": {
            "const": "bearer",
            "title": "Tokentype",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "actor",
          "expiresAt"
        ],
        "title": "DemoLoginResponse",
        "type": "object"
      },
...
      "FactGraph": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "title": "Facts",
            "type": "array"
          }
        },
        "required": [
          "facts"
        ],
        "title": "FactGraph",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManifestEvidence": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "chunk": {
            "$ref": "#/components/schemas/EvidenceChunk"
          }
        },
        "required": [
          "alias",
          "chunk"
        ],
        "title": "ManifestEvidence",
        "type": "object"
      },
...
      "NurseAction": {
        "additionalProperties": false,
        "properties": {
          "actionId": {
            "title": "Actionid",
            "type": "string"
          },
          "actionType": {
            "$ref": "#/components/schemas/ActionType"
          },
          "actorId": {
            "title": "Actorid",
            "type": "string"
          },
          "overrideDisposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Disposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "patientResponse": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "targetRecommendationId": {
            "title": "Targetrecommendationid",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "actionId",
          "targetRecommendationId",
          "actorId",
          "actionType",
          "rationale",
          "timestamp",
          "overrideDisposition",
          "patientResponse"
        ],
        "title": "NurseAction",
        "type": "object"
      },
...
      "PatientContext": {
        "additionalProperties": false,
        "properties": {
          "allergies": {
            "items": {
              "type": "string"
            },
            "title": "Allergies",
            "type": "array"
          },
          "demographics": {
            "$ref": "#/components/schemas/PatientDemographics"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "history": {
            "items": {
              "type": "string"
            },
            "title": "History",
            "type": "array"
          },
          "medications": {
            "items": {
              "type": "string"
            },
            "title": "Medications",
            "type": "array"
          },
          "patientId": {
            "title": "Patientid",
            "type": "string"
          },
          "synthetic": {
            "const": true,
            "title": "Synthetic",
            "type": "boolean"
          }
        },
        "required": [
          "patientId",
          "displayName",
          "demographics",
          "history",
          "medications",
          "allergies",
          "synthetic"
        ],
        "title": "PatientContext",
        "type": "object"
      },
      "PatientDemographics": {
        "additionalProperties": false,
        "properties": {
          "ageInMonths": {
            "maximum": 1560.0,
            "minimum": 0.0,
            "title": "Ageinmonths",
            "type": "integer"
          },
          "biologicalSex": {
            "$ref": "#/components/schemas/BiologicalSex"
          },
          "isPregnant": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ispregnant"
          }
        },
        "required": [
          "ageInMonths",
          "isPregnant",
          "biologicalSex"
        ],
        "title": "PatientDemographics",
        "type": "object"
      },
      "PatientResponse": {
        "additionalProperties": false,
        "properties": {
          "responseId": {
            "title": "Responseid",
            "type": "string"
          },
          "responseType": {
            "$ref": "#/components/schemas/PatientResponseType"
          },
          "statedPreference": {
            "title": "Statedpreference",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "responseId",
          "responseType",
          "statedPreference",
          "timestamp"
        ],
        "title": "PatientResponse",
        "type": "object"
      },
...
      "ReadinessState": {
        "enum": [
          "READY",
          "INCOMPLETE",
          "UNGROUNDED",
          "SAFETY_LOCKED"
        ],
        "title": "ReadinessState",
        "type": "string"
      },
...
      "SessionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "caseId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Caseid",
            "type": "string"
          }
        },
        "required": [
          "caseId"
        ],
        "title": "SessionCreateRequest",
        "type": "object"
      },
      "SessionPermissions": {
        "additionalProperties": false,
        "properties": {
          "authorizationServiceAvailable": {
            "title": "Authorizationserviceavailable",
            "type": "boolean"
          },
          "canAccept": {
            "title": "Canaccept",
            "type": "boolean"
          },
          "canDocumentRefusal": {
            "title": "Candocumentrefusal",
            "type": "boolean"
          },
          "canEscalate": {
            "title": "Canescalate",
            "type": "boolean"
          },
          "canOverride": {
            "title": "Canoverride",
            "type": "boolean"
          },
          "canView": {
            "title": "Canview",
            "type": "boolean"
          }
        },
        "required": [
          "canView",
          "canAccept",
          "canDocumentRefusal",
          "canEscalate",
          "canOverride",
          "authorizationServiceAvailable"
        ],
        "title": "SessionPermissions",
        "type": "object"
      },
      "SessionSummary": {
        "additionalProperties": false,
        "properties": {
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "patientName": {
            "title": "Patientname",
            "type": "string"
          },
          "readinessState": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "stateVersion": {
            "title": "Stateversion",
            "type": "integer"
          }
        },
        "required": [
          "sessionId",
          "patientName",
          "caseId",
          "stateVersion",
          "readinessState"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/NurseAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "careTeamId": {
            "title": "Careteamid",
            "type": "string"
          },
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "latestResponse": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "nurseNotes": {
            "title": "Nursenotes",
            "type": "string"
          },
          "originalTranscript": {
            "title": "Originaltranscript",
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientContext"
          },
          "patientOutcome": {
            "enum": [
              "UNRECORDED",
              "CARE_ACCEPTED",
              "REMAINS_HOME",
              "CLINICIAN_REVIEW"
            ],
            "title": "Patientoutcome",
            "type": "string"
          },
          "patientResponses": {
            "items": {
              "$ref": "#/components/schemas/PatientResponse"
            },
            "title": "Patientresponses",
            "type": "array"
          },
          "permissions": {
            "$ref": "#/components/schemas/SessionPermissions"
          },
          "recommendationStatus": {
            "enum": [
              "AWAITING_REVIEW",
              "ACCEPTED",
              "DECLINED_BY_PATIENT",
              "ESCALATED",
              "OVERRIDDEN"
            ],
            "title": "Recommendationstatus",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/SystemRecommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TriageState"
          }
        },
        "required": [
          "sessionId",
          "caseId",
          "careTeamId",
          "patient",
          "originalTranscript",
          "nurseNotes",
          "state",
          "latestResponse",
          "recommendations",
          "actions",
          "patientResponses",
          "recommendationStatus",
          "patientOutcome",
          "permissions"
        ],
        "title": "SessionView",
        "type": "object"
      },
...
      "SystemRecommendation": {
        "additionalProperties": false,
        "properties": {
          "computedStateVersion": {
            "title": "Computedstateversion",
            "type": "integer"
          },
          "readinessState": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "recommendationId": {
            "title": "Recommendationid",
            "type": "string"
          },
          "recommendedDisposition": {
            "$ref": "#/components/schemas/Disposition"
          },
          "safetyLocked": {
            "title": "Safetylocked",
            "type": "boolean"
          },
          "supersedesRecommendationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedesrecommendationid"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "validatedRationale": {
            "items": {
              "$ref": "#/components/schemas/RationaleClaim"
            },
            "title": "Validatedrationale",
            "type": "array"
          }
        },
        "required": [
          "recommendationId",
          "recommendedDisposition",
          "readinessState",
          "safetyLocked",
          "validatedRationale",
          "computedStateVersion",
          "timestamp",
          "supersedesRecommendationId"
        ],
        "title": "SystemRecommendation",
        "type": "object"
      },
      "TriageRequest": {
        "additionalProperties": false,
        "properties": {
          "baseStateVersion": {
            "minimum": 0.0,
            "title": "Basestateversion",
            "type": "integer"
          },
          "nurseNotes": {
            "anyOf": [
              {
                "maxLength": 6000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nursenotes"
          },
          "patientDemographics": {
            "$ref": "#/components/schemas/PatientDemographics"
          },
          "patientId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Patientid",
            "type": "string"
          },
          "sessionId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Sessionid",
            "type": "string"
          },
          "symptomDescription": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Symptomdescription",
            "type": "string"
          },
          "turnId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Turnid",
            "type": "string"
          },
          "vitalSigns": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalSigns"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "sessionId",
          "turnId",
          "baseStateVersion",
          "patientId",
          "symptomDescription",
          "patientDemographics"
        ],
        "title": "TriageRequest",
        "type": "object"
      },
      "TriageResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ManifestEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "factGraph": {
            "$ref": "#/components/schemas/FactGraph"
          },
          "missingInformation": {
            "items": {
              "type": "string"
            },
            "title": "Missinginformation",
            "type": "array"
          },
          "mode": {
            "enum": [
              "DETERMINISTIC",
              "PROVIDER",
              "MANUAL"
            ],
            "title": "Mode",
            "type": "string"
          },
          "rationale": {
            "items": {
              "$ref": "#/components/schemas/RationaleClaim"
            },
            "title": "Rationale",
            "type": "array"
          },
          "readinessState": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "recommendation": {
            "$ref": "#/components/schemas/SystemRecommendation"
          },
          "responseId": {
            "title": "Responseid",
            "type": "string"
          },
          "retrievalManifest": {
            "$ref": "#/components/schemas/RetrievalManifest"
          },
          "safety": {
            "$ref": "#/components/schemas/SafetyAssessment"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "stateVersion": {
            "title": "Stateversion",
            "type": "integer"
          },
          "systemAlerts": {
            "items": {
              "type": "string"
            },
            "title": "Systemalerts",
            "type": "array"
          },
          "timings": {
            "$ref": "#/components/schemas/StageTimings"
          },
          "turnId": {
            "title": "Turnid",
            "type": "string"
          },
          "validationCodes": {
            "items": {
              "type": "string"
            },
            "title": "Validationcodes",
            "type": "array"
          }
        },
        "required": [
          "responseId",
          "sessionId",
          "turnId",
          "stateVersion",
          "readinessState",
          "recommendation",
          "rationale",
          "evidence",
          "missingInformation",
          "systemAlerts",
          "factGraph",
          "safety",
          "retrievalManifest",
          "timings",
          "mode",
          "validationCodes"
        ],
        "title": "TriageResponse",
        "type": "object"
      },
      "TriageState": {
        "additionalProperties": false,
        "properties": {
          "activeTurnId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activeturnid"
          },
          "currentStateVersion": {
            "title": "Currentstateversion",
            "type": "integer"
          },
          "processingStatus": {
            "$ref": "#/components/schemas/ProcessingStatus"
          },
          "readinessState": {
            "$ref": "#/components/schemas/ReadinessState"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "processingStatus",
          "readinessState",
          "currentStateVersion",
          "activeTurnId"
        ],
        "title": "TriageState",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VitalSigns": {
        "additionalProperties": false,
        "properties": {
          "heartRate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 350.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartrate"
          },
          "oxygenSaturation": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oxygensaturation"
          },
          "respiratoryRate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 120.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Respiratoryrate"
          },
          "temperatureC": {
            "anyOf": [
              {
                "maximum": 45.0,
                "minimum": 25.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperaturec"
          }
        },
        "title": "VitalSigns",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Synthetic demonstration only. Not a medical device or real clinical guidance.",
    "title": "Meridian Patient Navigation & Triage Copilot",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/demo": {
      "post": {
        "operationId": "demoLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Demo Login"
      }
    },
...
    "/api/sessions": {
      "get": {
        "operationId": "listSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Listsessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions"
      },
      "post": {
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Session"
      }
    },
...
    "/api/triage": {
      "post": {
        "operationId": "runTriage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Triage"
      }
    }
  }
}